feat(foundation): always publish CDC functions using nested stacks
When all resources live in a single stack, it doesn't take a lot of
lambdas to rapidly use up the the resource quota for a template. I had
been trying to maintain both nested a "smooshed" styles, but I'm not
planning to use smooshed and it's a pain to maintain both. As the only
user, breaking changes are tolerable.

BREAKING CHANGE: All cdc functionaility will deploy in nested stacks.
This will likely require further template processing on the consumer
side to ensure all the appopriate values are applied.