fix: change condition name from "development" to "carpentry" to avoid side effects
Apparently, and unsurprisingly, many tools already use "development", but since ts is not quite
commonjs and not quite esm, none of those tools do the right thing with typical ts file. Long
run, the right thing to do seems to convert all to .mts, but that probably as its own set of
surprises.